+++ src/constants.js
@@ -1,0 +1,24 @@
+module.exports = {
+		//COMMAND IDs
+	CMD_SET_SEQ_TRANSPORT_MODE: 3,
+	CMD_MOVE_SEQ_TO_CUE: 4,
+	CMD_MOVE_SEQ_TO_LASTNEXTCUE: 7,
+	CMD_RESET_ALL: 9,
+	CMD_CLEAR_ALL_ACTIVE: 13,
+	CMD_TOGGLE_FULLSCREEN: 17,
+	CMD_STORE_ACTIVE: 25,
+	CMD_CLEAR_SELECTION: 48,
+	CMD_IGNORE_NEXT_CUE: 55,
+	CMD_SAVE_PROJECT: 62,
+	CMD_SET_SITE_IP: 71,
+	CMD_APPLY_VIEW: 103,
+	CMD_SET_SEQ_SELECTION: 299,
+	CMD_GET_SEQ_TRANSPORTMODE: 72,
+	CMD_GET_SEQ_TIME: 73,	
+	CMD_GET_REMAINING_TIME_UNTIL_NEXT_CUE: 78,		
+
+	feedbackstate: {
+		seqstate: 'Stop',
+		remainingQtime: 'Normal',
+	},
+}
